requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
imageio==2.31.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.2
llvmlite==0.40.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.8.0
pooch==1.7.0
protobuf==4.23.3
PyMatting==1.1.8
pyreadline3==3.4.1
PyWavelets==1.4.1
rembg==2.0.47
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.0
sympy==1.12
tifffile==2023.4.12
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==2.3.6
